Remove pagre re-render when switching preview mode
