+++ indico/modules/events/registration/controllers/management/regforms.py
@@ -22,6 +22,7 @@ from indico.modules.events.features.util import set_feature_enabled
 from indico.modules.events.models.events import EventType
 from indico.modules.events.payment import payment_settings
 from indico.modules.events.registration import logger, registration_settings
+from indico.modules.events.registration.controllers.display import ParticipantListMixin
 from indico.modules.events.registration.controllers.management import RHManageRegFormBase, RHManageRegFormsBase
 from indico.modules.events.registration.forms import (MultiFormsAnnouncementForm, ParticipantsDisplayForm,
                                                       ParticipantsDisplayFormColumnsForm, RegistrationFormCloneForm,
@@ -75,6 +76,33 @@ class RHParticipantListPreview(RHManageRegFormsBase):
 
     def _process(self):
         return self.view_class.render_template('display/participant_list.html', self.event, preview=self.preview)
+
+
+class RHParticipantListPreviewREST(ParticipantListMixin, RHManageRegFormsBase):
+    """REST API for the preview of the participant list."""
+
+    @use_kwargs(
+        {'guest': fields.Bool(load_default=False)}, location='query'
+    )
+    def _process_args(self, guest):
+        RHManageRegFormsBase._process_args(self)
+        self.preview = 'guest' if guest else 'participant'
+
+    def is_participant(self, user):
+        return self.preview == 'participant'
+
+    def _process(self):
+        is_participant = self.is_participant(session.user)
+        tables, merged, regforms = self._get_participant_list_tables(is_participant)
+
+        num_participants = sum(table['num_participants'] for table in tables)
+
+        return jsonify({
+            'published': bool(regforms),
+            'merged': merged,
+            'num_participants': num_participants,
+            'tables': tables
+        })
 
 
 class RHManageRegistrationFormsDisplay(RHManageRegFormsBase):
